+++ capreolus/tokenizer/base.py
@@ -1,0 +1,5 @@
+from capreolus import ModuleBase, Dependency, ConfigOption
+
+
+class Tokenizer(ModuleBase):
+    module_type = "tokenizer"
